Monday Fifth

+++ public/delete.php
@@ -1,8 +1,13 @@
-<?php
+<h2>Confirm deletion</h2>
+<form method="post">
+  <input type="submit" name="delete" value="Delete">
+  <input type="submit" name="cancel" value="Cancel">
+</form>
 
+<?php
 require "musician.php";
 
-if (isset($_GET["Id"])) {
+if (isset($_GET["Id"]) && isset($_POST["delete"])) {
   try {
     $newMusician = new Musician($_GET);
     $newMusician->deleteRecord();
@@ -11,5 +16,7 @@ if (isset($_GET["Id"])) {
   } catch(PDOException $error) {
 
   }
+} elseif (isset($_POST["cancel"])) {
+  header ('Location: index.php');
 }
 ?>
